refactor language_server tests and parser_library tests into more folders
