+++ language_server/test/send_message_provider_mock.h
@@ -15,7 +15,7 @@
 #ifndef HLASMPLUGIN_LANGUAGESERVER_TEST_SEND_MESSAGE_PROVIDER_MOCK_H
 #define HLASMPLUGIN_LANGUAGESERVER_TEST_SEND_MESSAGE_PROVIDER_MOCK_H
 
-#include "../src/server.h"
+#include "server.h"
 
 class send_message_provider_mock : public hlasm_plugin::language_server::send_message_provider
 {
