+++ language_server/test/server_test.h
@@ -19,11 +19,11 @@
 #include "json.hpp"
 #include "gmock/gmock.h"
 
-#include "../src/lsp/lsp_server.h"
+#include "lsp/lsp_server.h"
 #include "workspace_manager.h"
-#include "../src/lsp/feature_workspace_folders.h"
-#include "../src/lsp/feature_text_synchronization.h"
-#include "../src/feature.h"
+#include "lsp/feature_workspace_folders.h"
+#include "lsp/feature_text_synchronization.h"
+#include "feature.h"
 
 
 #include "send_message_provider_mock.h"
