+++ parser_library/test/checking/mach_instr_diag_test.cpp
@@ -14,7 +14,7 @@
 
 #include "gtest/gtest.h"
 
-#include "common_testing.h"
+#include "../common_testing.h"
 
 TEST(diagnostics, second_par_omitted)
 {
