+++ parser_library/test/context/org_test.cpp
@@ -14,7 +14,7 @@
 
 #include "gtest/gtest.h"
 
-#include "common_testing.h"
+#include "../common_testing.h"
 
 //tests for ORG instruction
 TEST(org, duplicate_labels)
