+++ parser_library/test/debugging/debug_event_consumer_s_mock.h
@@ -15,7 +15,7 @@
 #ifndef HLASMPLUGIN_PARSERLIBRARY_TEST_DEBUG_EVENT_CONSUMER_S_MOCK_H
 #define HLASMPLUGIN_PARSERLIBRARY_TEST_DEBUG_EVENT_CONSUMER_S_MOCK_H
 
-#include "../src/debugging/debugger.h"
+#include "debugging/debugger.h"
 
 using namespace hlasm_plugin::parser_library::debugging;
