+++ parser_library/test/parsing/parser_test.cpp
@@ -14,7 +14,7 @@
 
 #include "gtest/gtest.h"
 
-#include "common_testing.h"
+#include "../common_testing.h"
 #include "expressions/visitors/expression_evaluator.h"
 #include "parsing/parser_tools.h"
